nix: remove flake-parts
As far as abstraction libraries go, flake-parts is the best one out there. Except in this scenario, it serves no purpose. We are not in need of most of its features, and it complicates the schema for newcomers to Nix. Remove flake-parts and add a single function to abstract systems.