nix: add follows declarations to inputs of dependencies (#493)
This avoids there being two distinct instances of Nixpkgs and systems
in the dependency tree: one for walker and one for elephant. This also
helps users downstream, who otherwise have no easy way to pin
everything in the dependency graph to the same Nixpkgs, except by
hoisting all transitive dependencies as explicit inputs to their own
flakes. Without this, the number of Nixpkgs instances can grow
exponentially with the depth of the dependency graph.